factorize both files
