+++ ModelBundle/DataFixtures/MongoDB/LoadStatusData.php
@@ -2,18 +2,14 @@
 
 namespace OpenOrchestra\ModelBundle\DataFixtures\MongoDB;
 
-use Doctrine\Common\DataFixtures\AbstractFixture;
-use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
-use OpenOrchestra\ModelBundle\Document\Status;
-use OpenOrchestra\ModelBundle\Document\TranslatedValue;
 use OpenOrchestra\ModelInterface\DataFixtures\OrchestraProductionFixturesInterface;
 use OpenOrchestra\ModelInterface\DataFixtures\OrchestraFunctionalFixturesInterface;
 
 /**
  * Class LoadStatusData
  */
-class LoadStatusData extends AbstractFixture implements OrderedFixtureInterface, OrchestraProductionFixturesInterface, OrchestraFunctionalFixturesInterface
+class LoadStatusData extends AbstractLoadStatus implements OrchestraProductionFixturesInterface, OrchestraFunctionalFixturesInterface
 {
     /**
      * @param ObjectManager $manager
@@ -34,45 +30,5 @@ class LoadStatusData extends AbstractFixture implements OrderedFixtureInterface,
     public function getOrder()
     {
         return 40;
-    }
-
-    /**
-     * @param string $enName
-     * @param string $frName
-     * @param string $name
-     * @param string $color
-     * @param bool   $published
-     * @param bool   $initial
-     *
-     * @return Status
-     */
-    protected function loadStatus($enName, $frName, $name, $color, $published = false, $initial = false)
-    {
-        $value = new Status();
-        $value->setName($name);
-        $value->setPublished($published);
-        $value->setInitial($initial);
-        $value->addLabel($this->generateTranslatedValue('en', $enName));
-        $value->addLabel($this->generateTranslatedValue('fr', $frName));
-        $value->setDisplayColor($color);
-
-        $this->addReference('status-' . $name, $value);
-
-        return $value;
-    }
-
-    /**
-     * @param string $language
-     * @param string $value
-     *
-     * @return TranslatedValue
-     */
-    protected function generateTranslatedValue($language, $value)
-    {
-        $translatedValue = new TranslatedValue();
-        $translatedValue->setLanguage($language);
-        $translatedValue->setValue($value);
-
-        return $translatedValue;
     }
 }
