+++ ModelBundle/DataFixtures/MongoDB/LoadStatusDataFunctional.php
@@ -2,17 +2,13 @@
 
 namespace OpenOrchestra\ModelBundle\DataFixtures\MongoDB;
 
-use Doctrine\Common\DataFixtures\AbstractFixture;
-use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
-use OpenOrchestra\ModelBundle\Document\Status;
-use OpenOrchestra\ModelBundle\Document\TranslatedValue;
 use OpenOrchestra\ModelInterface\DataFixtures\OrchestraFunctionalFixturesInterface;
 
 /**
  * Class LoadStatusDataFunctional
  */
-class LoadStatusDataFunctional extends AbstractFixture implements OrderedFixtureInterface, OrchestraFunctionalFixturesInterface
+class LoadStatusDataFunctional extends AbstractLoadStatus implements OrchestraFunctionalFixturesInterface
 {
     /**
      * @param ObjectManager $manager
@@ -32,45 +28,5 @@ class LoadStatusDataFunctional extends AbstractFixture implements OrderedFixture
     public function getOrder()
     {
         return 45;
-    }
-
-    /**
-     * @param string $enName
-     * @param string $frName
-     * @param string $name
-     * @param string $color
-     * @param bool   $published
-     * @param bool   $initial
-     *
-     * @return Status
-     */
-    protected function loadStatus($enName, $frName, $name, $color, $published = false, $initial = false)
-    {
-        $value = new Status();
-        $value->setName($name);
-        $value->setPublished($published);
-        $value->setInitial($initial);
-        $value->addLabel($this->generateTranslatedValue('en', $enName));
-        $value->addLabel($this->generateTranslatedValue('fr', $frName));
-        $value->setDisplayColor($color);
-
-        $this->addReference('status-' . $name, $value);
-
-        return $value;
-    }
-
-    /**
-     * @param string $language
-     * @param string $value
-     *
-     * @return TranslatedValue
-     */
-    protected function generateTranslatedValue($language, $value)
-    {
-        $translatedValue = new TranslatedValue();
-        $translatedValue->setLanguage($language);
-        $translatedValue->setValue($value);
-
-        return $translatedValue;
     }
 }
